Fix startup issue.
A total hack.  For reasons as of yet unknown, when the ambilight program starts instantly on boot on the Pi, it now fails to read the video stream properly and the colors sent to the leds are in the wrong range.  Delaying ensures that all necessary subsystems are initialized before trying to read the video stream (I guess?)